Update animation and add CardList component to manage cards in a list
